+++ CHANGELOG.md
@@ -13,6 +13,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 -->
 
+### Added
+- New descriptions and ordering for project templates (MAD-1428)
+- New "Duplicate project" action on projects
+
+
+### Fixed
+- Fixed styling for Simple HTML Block (reset by tailwind / MAD-1449)
+- Fixed incorrect button label on Manifest page (MAD-1447)
+- Fixed blocking of project contribution in Manifest Granularity (MAD-1447)
+- Fixed bug with correcting submissions when there are multiple choices (MAD-1453)
+- Fixed project dashboard default option (MAD-1452)
+- Fixed labeling of "Export project"
+
+### Removed
+- Removed config option `projectPageOptions.reviewerDashboard` (MAD-1452)
+
 ## [v2.2.1](https://github.com/digirati-co-uk/madoc-platform/compare/v2.2.0...v2.2.1)